Add reusable principal combobox #1208
-Refactored RichComboBox - added parent class (BaseRichComboBox) that allows it's loader to load items of one type and then convert them (loadedItemToDisplayValue) to another one
-Added builders for RichComboBox inheritors
-Added PrincipalContainerCombobox to allow comboboxes with principal access reuse common code  (principal loader, principal options viewer, value changed listeners), just need to extend PrincipalContainer
